+++ test/operations/legalise_edge.jl
@@ -14,7 +14,7 @@ include("../helper_functions.jl")
     i, j, r = 5, 1, 7
     e = DT.get_adjacent(tri, j, i)
     DT.legalise_edge!(tri, i, j, r)
-    true_T = Set{NTuple{3,Int64}}([
+    true_T = Set{NTuple{3,Int}}([
         (3, 2, 5),
         (1, 3, 7),
         (3, 5, 7),
@@ -41,14 +41,14 @@ include("../helper_functions.jl")
         )
     )
     true_adj2v = Dict(
-        DT.BoundaryIndex => Set{NTuple{2,Int64}}([(4, 5), (5, 2), (2, 6), (6, 4)]),
-        1 => Set{NTuple{2,Int64}}([(6, 3), (3, 7), (7, 4), (4, 6)]),
-        2 => Set{NTuple{2,Int64}}([(5, 3), (3, 6)]),
-        3 => Set{NTuple{2,Int64}}([(2, 5), (5, 7), (7, 1), (1, 6), (6, 2)]),
-        4 => Set{NTuple{2,Int64}}([(6, 1), (1, 7), (7, 5)]),
-        5 => Set{NTuple{2,Int64}}([(4, 7), (7, 3), (3, 2)]),
-        6 => Set{NTuple{2,Int64}}([(2, 3), (3, 1), (1, 4)]),
-        7 => Set{NTuple{2,Int64}}([(3, 5), (5, 4), (4, 1), (1, 3)])
+        DT.BoundaryIndex => Set{NTuple{2,Int}}([(4, 5), (5, 2), (2, 6), (6, 4)]),
+        1 => Set{NTuple{2,Int}}([(6, 3), (3, 7), (7, 4), (4, 6)]),
+        2 => Set{NTuple{2,Int}}([(5, 3), (3, 6)]),
+        3 => Set{NTuple{2,Int}}([(2, 5), (5, 7), (7, 1), (1, 6), (6, 2)]),
+        4 => Set{NTuple{2,Int}}([(6, 1), (1, 7), (7, 5)]),
+        5 => Set{NTuple{2,Int}}([(4, 7), (7, 3), (3, 2)]),
+        6 => Set{NTuple{2,Int}}([(2, 3), (3, 1), (1, 4)]),
+        7 => Set{NTuple{2,Int}}([(3, 5), (5, 4), (4, 1), (1, 3)])
     )
     true_DG = relabel(UndirectedGraph([
             0 0 1 0 1 1 1 0
